Changes
Please notice that I change all the files extension to .php so that we can process php code and html code in the same file.

The login system is completed and working perfectly, if you notice any mistake, please correct me.

TODO -> Finishing research progress part.
TODO -> Create a profile for the users.